Ported to Idris dev tip, requiring adding --yes-really.
Introduced a universive of foreign Java types. This, in particular,
allows dynamic Java classes to be allocated (i.e. new) and method
instances to be called. Also provides the basis for a type promoter
that generates an interface for class.

Added some basic tests, although not automatically running them yet.